chore: port the authz example to use the Python SDK
We also update the requirements.txt file (to say we depend on the SDK).
In order to specify some fully working versions within it
(which is typical of a requirements.txt file), there's also now a
distinction between a requirements.in file (which lists the dependencies)
and the requirements.txt file (which includes transitive dependencies and
versions of them that are known to work). This is parallel to a package.json
and lockfile in Javascript land, but in Python the common tool is:

    https://pypi.org/project/pip-tools/

Otherwise, this really preserves most of the structure of the file and
just moves the manual calls to requests to instead use the library.

A subsequent commit will do some minor tidying.